requirements.txt
tweepy
vrchatapi
python-dotenv